#55: moved argument function to assert
- added temporary BC layer in core

----
Move argument assertion function from "core" to "assert" package
https://github.com/monumentjs/core/issues/55
