+++ packages/assert/index.ts
@@ -1,6 +1,7 @@
 // Export public API
 
 export * from './src/assert';
+export * from './src/argument';
 
 // backward compatibility layer
 export { AssertionException } from '@monument/exceptions';
