+++ packages/date/package.json
@@ -26,6 +26,7 @@
     "access": "public"
   },
   "dependencies": {
+    "@monument/assert": "^0.15.0",
     "@monument/core": "^0.15.0",
     "@monument/exceptions": "^0.15.0",
     "@monument/collections": "^0.15.0",
