+++ packages/text/package.json
@@ -25,6 +25,7 @@
     }
   ],
   "dependencies": {
+    "@monument/assert": "^0.15.0",
     "@monument/core": "^0.15.0",
     "@monument/collections": "^0.15.0",
     "@monument/exceptions": "^0.15.0"
