+++ packages/version/package.json
@@ -26,6 +26,7 @@
     }
   ],
   "dependencies": {
+    "@monument/assert": "^0.15.0",
     "@monument/core": "^0.15.0",
     "@monument/exceptions": "^0.15.0",
     "@monument/text": "^0.15.0"
