+++ packages/version/src/VersionBuilder.ts
@@ -1,4 +1,5 @@
-import { argument, Builder } from '@monument/core';
+import { argument } from '@monument/assert';
+import { Builder } from '@monument/core';
 import { Version } from './Version';
 import { ReleaseStatus } from './ReleaseStatus';
 import { VersionComponents } from './VersionComponents';
